Move tracks to unused. Fix scss deprecation warnings.
