+++ assets/js/components/tracks/annotation_tracks.ts
@@ -1,4 +1,4 @@
-import { removeChildren } from "../../track/utils";
+import { removeChildren } from "../../unused/track/utils";
 import { BandTrack } from "./band_track";
 
 const template = document.createElement("template");
