+++ assets/js/components/tracks/band_track.ts
@@ -1,5 +1,5 @@
-import { getOverlapInfo, getTrackHeight } from "../../track/expand_track_utils";
-import { getBandYScale } from "../../track/utils";
+import { getOverlapInfo, getTrackHeight } from "../../expand_track_utils";
+import { getBandYScale } from "../../unused/track/utils";
 import { STYLE } from "../../util/constants";
 import { CanvasTrack } from "./canvas_track";
 import {
