+++ assets/js/components/tracks/render_utils.ts
@@ -1,4 +1,4 @@
-import { rangeSize } from "../../track/utils";
+import { rangeSize } from "../../unused/track/utils";
 import { STYLE } from "../../util/constants";
 
 export function renderBackground(
